Implement functions to evaluate a math expression
Implement a function that will convert a mathematical expression into
a postfix expression where operands come first then operations, And
implement another function that take a postfix expression as argument
and return the result. The use of postfix expression is to facilitate
the parsing and evaluation of the user math input.